Issue 18: move tests to apropriate case
